add IntegrityError check, display table

+++ app/check_functions.py
@@ -8,26 +8,26 @@ from app import db
 from werkzeug.utils import secure_filename
 
 UPLOAD_FOLDER = '/home/e.sergeev/private_repos/city-city/app/static/filestorage/'
-ALLOWED_EXTENSIONS = set(['pdf', 'jpeg', 'doc', 'docx', 'txt'])
+ALLOWED_EXTENSIONS = set(['pdf', 'jpeg', 'doc', 'docx', 'txt', 'jpg'])
 MAX_FILE_SIZE = 200 * 1024 * 1024
 
 def save_request_data(request):
     try:
         file = request.files['userfile']
+        filename = secure_filename(file.filename)
         if not(file):
             return 'You send no file'
         if not(check_file_size(file)):
             return 'File size is too big'
         elif not(check_file_extension(file.filename)):
             return 'Unsupported file extension'
-        elif regexp(request.form['article_name'], r'[^a-zA-Z\s]'):
+        elif regexp(request.form['article_name'], r'[^a-zA-Z\s0-9]'):
             return 'Wrong article_name'
-        elif regexp(request.form['author_name'], r'[^a-zA-Z\s]'):
+        elif regexp(request.form['author_name'], r'[^a-zA-Z\s0-9]'):
             return('wrong author_name')
-        elif regexp(request.form['email'], r'@\w+'):#do not check. why?
+        elif not(regexp(request.form['email'], r'[@\w+]')):#do not check. why?
             return('wrong email')
         else:
-            filename = secure_filename(file.filename)
             record = Articles(article_name = request.form['article_name'],\
                                      author_name = request.form['author_name'],\
                                      file_name = filename,\
@@ -39,12 +39,14 @@ def save_request_data(request):
             db.session.commit()
             return 'Succes'
     except exc.SQLAlchemyError as ex:
-        print ex
+        if ex.orig.pgcode == '23505':
+            return ex.orig.message.split('DETAIL:')[1]
+        print 'database exception', ex
         os.remove(os.path.join(UPLOAD_FOLDER, filename))
         db.session.rollback()
         return 'Enternal server Error, write system administrator'
     except Exception as ex:
-        print ex
+        print 'Exception', ex, type(ex)
         return 'Enternal server Error, write system administrator'
 
 def check_file_size(file):
